added data.py and transfered get_raw_data from preprocessor to data
